+++ route_prefix.py
@@ -16,3 +16,4 @@ PREFIX_LABELING = PREFIX + "/labeling"
 PREFIX_RECORD_IDE = PREFIX + "/record-ide"
 PREFIX_RECORD = PREFIX + "/record"
 PREFIX_WEAK_SUPERVISION = PREFIX + "/weak-supervision"
+PREFIX_LABELING_TASKS = PREFIX + "/labeling-tasks"
